Workshop 4 Winter Term 2021
Program that implements a simple validation on a series of user input values that are stored to arrays and later analyzed to produce a variety of summary reports. The program will ask for the user’s monthly income and then ask for the price and priority of a series of items the user would like to purchase in the future. It will store this information and allow the user to view predictions on how long it will take to save enough money to purchase their wish list items.